+++ Rollbar/RollbarTelemetryType.h
@@ -2,7 +2,9 @@
 
 #import <Foundation/Foundation.h>
 
-typedef enum {
+#pragma mark - RollbarTelemetryType
+
+typedef NS_ENUM(NSUInteger, RollbarTelemetryType) {
     RollbarTelemetryLog,
     RollbarTelemetryView,
     RollbarTelemetryError,
@@ -10,6 +12,30 @@ typedef enum {
     RollbarTelemetryNetwork,
     RollbarTelemetryConnectivity,
     RollbarTelemetryManual
-} RollbarTelemetryType;
+};
 
-NSString* RollbarStringFromTelemetryType(RollbarTelemetryType type);
+
+#pragma mark - RollbarLevel utility
+
+NS_ASSUME_NONNULL_BEGIN
+
+/// RollbarTelemetryType utility
+@interface RollbarTelemetryTypeUtil : NSObject
+
+/// Converts RollbarTelemetryType enum value to its string equivalent or default string.
+/// @param value RollbarTelemetryType enum value
++ (NSString *) RollbarTelemetryTypeToString:(RollbarTelemetryType)value;
+
+/// Converts string value into its  RollbarTelemetryType enum value equivalent or default enum value.
+/// @param value input string
++ (RollbarTelemetryType) RollbarTelemetryTypeFromString:(NSString *)value;
+
+@end
+
+NS_ASSUME_NONNULL_END
+
+#pragma mark - deprecated
+
+NSString* _Nonnull RollbarStringFromTelemetryType(RollbarTelemetryType type)
+DEPRECATED_MSG_ATTRIBUTE("Use [RollbarTelemetryTypeUtil RollbarTelemetryTypeToString:...] methods instead.");
+