+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The change log has moved to this repo's [GitHub Releases Page](https://github.co
 ## Release Notes
 
 **1.11.7**
+- refactor: resolve #256: Refactor RollbarTelemetry implementation.
 - chore: ref #255: Mark deprecated public API with proper deprecated attribute.
 
 **1.11.6**